Adapt the program for my one usage
It seems that all I need for this is to change some symlinks here and
there, and in some cases (e.g. alacritty) touch the main config file to
make them reload it. That can all be specified in a config file, which
is now the case (in ~/.config/theme-switcher.toml).

I also forced the update of the colorscheme at startup, because the
colorscheme is automatically updated depending on time of day on my
system, so it might not be the same as it was when I stopped my session
last. Incidentally, that made me adopt ashpd rather than raw zbus
(easier to understand the APIs for a dbus noob)